+++ src/graphics/GPUFilter.h
@@ -60,8 +60,8 @@ public:
 protected:
     void draw(GLTexturePtr pTex);
     const std::string& getStdShaderCode() const;
-    int getBlurKernelRadius(double stdDev) const;
-    GLTexturePtr calcBlurKernelTex(double stdDev, double opacity=1) const;
+    int getBlurKernelRadius(float stdDev) const;
+    GLTexturePtr calcBlurKernelTex(float stdDev, float opacity=1) const;
 
 private:
     PixelFormat m_PFSrc;
