+++ src/graphics/Pixel32.h
@@ -50,7 +50,7 @@ public:
     unsigned char getB() const;
     unsigned char getA() const;
     void flipRB();
-    void toHSL(double& h, double& s, double& l);
+    void toHSL(float& h, float& s, float& l);
 
     bool operator ==(const Pixel32 pix) const;
     bool operator !=(const Pixel32 pix) const;
