+++ src/player/LineNode.h
@@ -43,19 +43,19 @@ class AVG_API LineNode : public VectorNode
         const glm::vec2& getPos2() const;
         void setPos2(const glm::vec2& pt);
 
-        double getTexCoord1() const;
-        void setTexCoord1(double tc);
+        float getTexCoord1() const;
+        void setTexCoord1(float tc);
 
-        double getTexCoord2() const;
-        void setTexCoord2(double tc);
+        float getTexCoord2() const;
+        void setTexCoord2(float tc);
 
         virtual void calcVertexes(VertexArrayPtr& pVertexArray, Pixel32 color);
 
     private:
         glm::vec2 m_P1;
         glm::vec2 m_P2;
-        double m_TC1;
-        double m_TC2;
+        float m_TC1;
+        float m_TC2;
 };
 
 }
